backend/requirements.txt
django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5
Pillow>=10.0
python-dotenv>=1.0
twilio>=8.12
pytesseract>=0.3
imagehash>=4.3
requests>=2.31
